multi release jar removed
