+++ src/test/java/org/eclipse/yasson/records/CarWithCreator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -15,11 +15,11 @@ package org.eclipse.yasson.records;
 import jakarta.json.bind.annotation.JsonbCreator;
 import jakarta.json.bind.annotation.JsonbProperty;
 
-public record CarWithMultipleConstructorsAndCreator(String type, String color) {
+public record CarWithCreator(String type, String color) {
 
     @JsonbCreator
-    public CarWithMultipleConstructorsAndCreator(@JsonbProperty("type") String type) {
-        this(type, "red");
+    public static CarWithCreator create(@JsonbProperty("type") String type, @JsonbProperty("color") String color) {
+        return new CarWithCreator(type, color);
     }
 
 }
