+++ src/test/java/org/eclipse/yasson/records/CarWithDefaultConstructor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,10 +12,10 @@
 
 package org.eclipse.yasson.records;
 
-public record CarWithExtraMethod(String type, String color) {
+public record CarWithDefaultConstructor(String type, String color) {
 
-    public String type() {
-        return type;
+    public CarWithDefaultConstructor() {
+        this("some", "other");
     }
 
 }
