+++ src/test/java/org/eclipse/yasson/records/CarWithExtraMethod.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,14 +12,10 @@
 
 package org.eclipse.yasson.records;
 
-import jakarta.json.bind.annotation.JsonbCreator;
-import jakarta.json.bind.annotation.JsonbProperty;
+public record CarWithExtraMethod(String type, String color) {
 
-public record CarWithCreator(String type, String color) {
-
-    @JsonbCreator
-    public static CarWithCreator create(@JsonbProperty("type") String type, @JsonbProperty("color") String color) {
-        return new CarWithCreator(type, color);
+    public String type() {
+        return type;
     }
 
 }
