+++ src/test/java/org/eclipse/yasson/records/CarWithMultipleConstructorsAndCreator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,7 +12,14 @@
 
 package org.eclipse.yasson.records;
 
+import jakarta.json.bind.annotation.JsonbCreator;
 import jakarta.json.bind.annotation.JsonbProperty;
 
-public record Car(@JsonbProperty("typeChanged") String type, @JsonbProperty("colorChanged") String color) {
+public record CarWithMultipleConstructorsAndCreator(String type, String color) {
+
+    @JsonbCreator
+    public CarWithMultipleConstructorsAndCreator(@JsonbProperty("type") String type) {
+        this(type, "red");
+    }
+
 }
